+++ resources/src/components/Address.vue
@@ -1,63 +1,78 @@
 <template>
 	<div :class="$style.grid">
-		<Input
-			:label="labels.fullAddress"
-			:value="value.address"
-			@input="onInput('fullAddress', $event)"
-		/>
+		<div :class="$style.full">
+			<Input
+				:label="labels.fullAddress"
+				:value="value.address"
+				@input="onInput('fullAddress', $event)"
+				:disabled="hide"
+			/>
 
-		<Input
-			:label="labels.number"
-			:value="value.parts.number"
-			@input="onInput('number', $event)"
-		/>
+			<button
+				class="btn"
+				@click="onClear()"
+				type="button"
+			>
+				Clear
+			</button>
+		</div>
 
-		<Input
-			:label="labels.address"
-			:value="value.parts.address"
-			@input="onInput('address', $event)"
-		/>
+		<Fragment v-if="!hide">
+			<Input
+				:label="labels.number"
+				:value="value.parts.number"
+				@input="onInput('number', $event)"
+			/>
 
-		<Input
-			:label="labels.city"
-			:value="value.parts.city"
-			@input="onInput('city', $event)"
-		/>
+			<Input
+				:label="labels.address"
+				:value="value.parts.address"
+				@input="onInput('address', $event)"
+			/>
 
-		<Input
-			:label="labels.postcode"
-			:value="value.parts.postcode"
-			@input="onInput('postcode', $event)"
-		/>
+			<Input
+				:label="labels.city"
+				:value="value.parts.city"
+				@input="onInput('city', $event)"
+			/>
 
-		<Input
-			:label="labels.county"
-			:value="value.parts.county"
-			@input="onInput('county', $event)"
-		/>
+			<Input
+				:label="labels.postcode"
+				:value="value.parts.postcode"
+				@input="onInput('postcode', $event)"
+			/>
 
-		<Input
-			:label="labels.state"
-			:value="value.parts.state"
-			@input="onInput('state', $event)"
-		/>
+			<Input
+				:label="labels.county"
+				:value="value.parts.county"
+				@input="onInput('county', $event)"
+			/>
 
-		<Input
-			:label="labels.country"
-			:value="value.parts.country"
-			@input="onInput('country', $event)"
-		/>
+			<Input
+				:label="labels.state"
+				:value="value.parts.state"
+				@input="onInput('state', $event)"
+			/>
+
+			<Input
+				:label="labels.country"
+				:value="value.parts.country"
+				@input="onInput('country', $event)"
+			/>
+		</Fragment>
 	</div>
 </template>
 
 <script lang="js">
 	import { Component, Vue } from 'vue-property-decorator';
 	import Input from './Input';
+	import Fragment from './Fragment';
 	import { t } from '../filters/craft';
 
 	@Component({
 		components: {
 			Input,
+			Fragment,
 		},
 		props: {
 			value: {
@@ -70,6 +85,7 @@
 				},
 			},
 			fullAddressDirty: Boolean,
+			hide: Boolean,
 		}
 	})
 	export default class Address extends Vue {
@@ -98,6 +114,10 @@
 			});
 		}
 
+		onClear () {
+			this.$emit('clear');
+		}
+
 	}
 </script>
 
@@ -117,13 +137,27 @@
 			grid-template-columns: 1fr;
 		}
 
-		label:first-child,
+		div:first-child,
 		label:last-child:nth-child(even) {
 			grid-column: span 2;
 
 			@media only screen and (max-width: 998px) {
 				grid-column: span 1;
 			}
+		}
+	}
+
+	.full {
+		display: flex;
+		align-items: flex-end;
+
+		label {
+			width: 100%;
+			margin-right: 14px;
+		}
+
+		button {
+			font-size: 14px;
 		}
 	}
 </style>