+++ resources/src/components/Fragment.vue
@@ -1,0 +1,10 @@
+<script>
+	const Fragment = {
+		functional: true,
+		render (h, ctx) {
+			return ctx.children;
+		}
+	};
+
+	export default Fragment;
+</script>